requirement.txt
alembic==1.13.2
annotated-types==0.6.0
anyio==4.2.0
bcrypt==3.2.2
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
colorama==0.4.6
dnspython==2.5.0
ecdsa==0.19.0
email-validator==2.1.0.post1
fastapi==0.109.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
Mako==1.3.5
MarkupSafe==2.1.5
orjson==3.9.13
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
psycopg2==2.9.9
pyasn1==0.6.0
pycparser==2.22
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyJWT==2.8.0
pytest==8.2.2
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
PyYAML==6.0.1
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.31
starlette==0.36.3
typing_extensions==4.9.0
ujson==5.9.0
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
